+++ src/components/core/inputs/Checkbox/Checkbox-types.ts
@@ -3,6 +3,7 @@ export interface ContainerProps {
   label: string | React.ReactNode;
   selected: boolean;
   onChange: (name: string, selected: boolean) => void;
+  onBlur: (name: string, selected: boolean) => void;
   error: boolean;
   required?: boolean;
   disabled?: boolean;
@@ -12,4 +13,5 @@ export interface ContainerProps {
 export interface ViewProps extends ContainerProps {
   className: string;
   onChange: () => void;
+  onBlur: () => void;
 }