+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
 
 ### 👷 Changed
 
+- Change Checkbox (input) core component
+  - Add onBlur property
 - Change useForm hook
   - Add support for dropdown fields
   - Add onBlur property to checkbox fields